add validation for onchange func
